requirements.txt
Django==3.0.8
pytz==2019.3
gunicorn==20.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-ckeditor==5.9.0
psycopg2-binary==2.8.4
dj-database-url==0.5.0